[REG-2000] Implement proper test exposure patterns in the SDK
