+++ src/RGUnityBots/Assets/RegressionGames/Resources/BotSequences.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8b555841a17587b49902368dc616df42
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
