+++ src/RGUnityBots/Assets/RegressionGames/Resources/BotSequences/Sequence-Editor-Description.json
@@ -1,0 +1,6 @@
+{
+"name":"Sequence Editor Description",
+"description":"Sequence Editor Description",
+"segments":[
+
+]}
