+++ src/RGUnityBots/Assets/RegressionGames/Resources/BotSequences/Sequence-Editor-Description.json.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d76bd9b712a98c74a86cc4b8d17a11f0
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
