+++ src/RGUnityBots/Assets/Resources/RGInputSettingsCopy.txt
@@ -1,0 +1,1 @@
+{"InputManager":{"serializedVersion":"2","m_Axes":[{"serializedVersion":"3","m_Name":"Horizontal","descriptiveName":"","descriptiveNegativeName":"","negativeButton":"left","positiveButton":"right","altNegativeButton":"a","altPositiveButton":"d","gravity":3.0,"dead":0.0010000000474974514,"sensitivity":3.0,"snap":true,"invert":false,"type":0,"axis":0,"joyNum":0},{"serializedVersion":"3","m_Name":"Vertical","descriptiveName":"","descriptiveNegativeName":"","negativeButton":"down","positiveButton":"up","altNegativeButton":"s","altPositiveButton":"w","gravity":3.0,"dead":0.0010000000474974514,"sensitivity":3.0,"snap":true,"invert":false,"type":0,"axis":0,"joyNum":0},{"serializedVersion":"3","m_Name":"Fire1","descriptiveName":"","descriptiveNegativeName":"","negativeButton":"","positiveButton":"left ctrl","altNegativeButton":"","altPositiveButton":"mouse 0","gravity":1000.0,"dead":0.0010000000474974514,"sensitivity":1000.0,"snap":false,"invert":false,"type":0,"axis":0,"joyNum":0},{"serializedVersion":"3","m_Name":"Fire2","descriptiveName":"","descriptiveNegativeName":"","negativeButton":"","positiveButton":"left alt","altNegativeButton":"","altPositiveButton":"mouse 1","gravity":1000.0,"dead":0.0010000000474974514,"sensitivity":1000.0,"snap":false,"invert":false,"type":0,"axis":0,"joyNum":0},{"serializedVersion":"3","m_Name":"Fire3","descriptiveName":"","descriptiveNegativeName":"","negativeButton":"","positiveButton":"left shift","altNegativeButton":"","altPositiveButton":"mouse 2","gravity":1000.0,"dead":0.0010000000474974514,"sensitivity":1000.0,"snap":false,"invert":false,"type":0,"axis":0,"joyNum":0},{"serializedVersion":"3","m_Name":"Jump","descriptiveName":"","descriptiveNegativeName":"","negativeButton":"","positiveButton":"space","altNegativeButton":"","altPositiveButton":"","gravity":1000.0,"dead":0.0010000000474974514,"sensitivity":1000.0,"snap":false,"invert":false,"type":0,"axis":0,"joyNum":0},{"serializedVersion":"3","m_Name":"Mouse X","descriptiveName":"","descriptiveNegativeName":"","negativeButton":"","positiveButton":"","altNegativeButton":"","altPositiveButton":"","gravity":0.0,"dead":0.0,"sensitivity":0.10000000149011612,"snap":false,"invert":false,"type":1,"axis":0,"joyNum":0},{"serializedVersion":"3","m_Name":"Mouse Y","descriptiveName":"","descriptiveNegativeName":"","negativeButton":"","positiveButton":"","altNegativeButton":"","altPositiveButton":"","gravity":0.0,"dead":0.0,"sensitivity":0.10000000149011612,"snap":false,"invert":false,"type":1,"axis":1,"joyNum":0},{"serializedVersion":"3","m_Name":"Mouse ScrollWheel","descriptiveName":"","descriptiveNegativeName":"","negativeButton":"","positiveButton":"","altNegativeButton":"","altPositiveButton":"","gravity":0.0,"dead":0.0,"sensitivity":0.10000000149011612,"snap":false,"invert":false,"type":1,"axis":2,"joyNum":0},{"serializedVersion":"3","m_Name":"Horizontal","descriptiveName":"","descriptiveNegativeName":"","negativeButton":"","positiveButton":"","altNegativeButton":"","altPositiveButton":"","gravity":0.0,"dead":0.1899999976158142,"sensitivity":1.0,"snap":false,"invert":false,"type":2,"axis":0,"joyNum":0},{"serializedVersion":"3","m_Name":"Vertical","descriptiveName":"","descriptiveNegativeName":"","negativeButton":"","positiveButton":"","altNegativeButton":"","altPositiveButton":"","gravity":0.0,"dead":0.1899999976158142,"sensitivity":1.0,"snap":false,"invert":true,"type":2,"axis":1,"joyNum":0},{"serializedVersion":"3","m_Name":"Fire1","descriptiveName":"","descriptiveNegativeName":"","negativeButton":"","positiveButton":"joystick button 0","altNegativeButton":"","altPositiveButton":"","gravity":1000.0,"dead":0.0010000000474974514,"sensitivity":1000.0,"snap":false,"invert":false,"type":0,"axis":0,"joyNum":0},{"serializedVersion":"3","m_Name":"Fire2","descriptiveName":"","descriptiveNegativeName":"","negativeButton":"","positiveButton":"joystick button 1","altNegativeButton":"","altPositiveButton":"","gravity":1000.0,"dead":0.0010000000474974514,"sensitivity":1000.0,"snap":false,"invert":false,"type":0,"axis":0,"joyNum":0},{"serializedVersion":"3","m_Name":"Fire3","descriptiveName":"","descriptiveNegativeName":"","negativeButton":"","positiveButton":"joystick button 2","altNegativeButton":"","altPositiveButton":"","gravity":1000.0,"dead":0.0010000000474974514,"sensitivity":1000.0,"snap":false,"invert":false,"type":0,"axis":0,"joyNum":0},{"serializedVersion":"3","m_Name":"Jump","descriptiveName":"","descriptiveNegativeName":"","negativeButton":"","positiveButton":"joystick button 3","altNegativeButton":"","altPositiveButton":"","gravity":1000.0,"dead":0.0010000000474974514,"sensitivity":1000.0,"snap":false,"invert":false,"type":0,"axis":0,"joyNum":0},{"serializedVersion":"3","m_Name":"Submit","descriptiveName":"","descriptiveNegativeName":"","negativeButton":"","positiveButton":"return","altNegativeButton":"","altPositiveButton":"joystick button 0","gravity":1000.0,"dead":0.0010000000474974514,"sensitivity":1000.0,"snap":false,"invert":false,"type":0,"axis":0,"joyNum":0},{"serializedVersion":"3","m_Name":"Submit","descriptiveName":"","descriptiveNegativeName":"","negativeButton":"","positiveButton":"enter","altNegativeButton":"","altPositiveButton":"space","gravity":1000.0,"dead":0.0010000000474974514,"sensitivity":1000.0,"snap":false,"invert":false,"type":0,"axis":0,"joyNum":0},{"serializedVersion":"3","m_Name":"Cancel","descriptiveName":"","descriptiveNegativeName":"","negativeButton":"","positiveButton":"escape","altNegativeButton":"","altPositiveButton":"joystick button 1","gravity":1000.0,"dead":0.0010000000474974514,"sensitivity":1000.0,"snap":false,"invert":false,"type":0,"axis":0,"joyNum":0}],"m_UsePhysicalKeys":false}}
