+++ src/RGUnityBots/Assets/Resources/RGInputSettingsCopy.txt.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: acc031d0225778c41984587e3502e536
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
