+++ src/RGUnityBots/Packages/manifest.json
@@ -44,5 +44,6 @@
     "com.unity.modules.vr": "1.0.0",
     "com.unity.modules.wind": "1.0.0",
     "com.unity.modules.xr": "1.0.0"
-  }
+  },
+  "testables": ["gg.regression.unity.bots"]
 }
