+++ src/gg.regression.unity.bots/Tests/Scripts/LegacyInput/RGLegacyInputUtilityTests.cs
@@ -2,9 +2,10 @@
 using System.Collections.Generic;
 using NUnit.Framework;
 using RegressionGames.RGLegacyInputUtility;
+using RegressionGames.TestFramework;
 using UnityEngine;
 
-namespace Tests.TestFramework.LegacyInput
+namespace RegressionGames.Tests.LegacyInput
 {
     [TestFixture]
     public class RGLegacyInputUtilityTests
