+++ src/gg.regression.unity.bots/Tests/Scripts/ObjectDetection.meta
