+++ src/gg.regression.unity.bots/Tests/Scripts/ObjectDetection/CVObjectDetectionQueryTests.cs.meta
