+++ src/gg.regression.unity.bots/Tests/Scripts/RGOverlay/RGCreateNewSequenceButton.test.cs.meta
