+++ src/gg.regression.unity.bots/Tests/Scripts/RGOverlay/RGSequenceEntry.test.cs.meta
