+++ src/gg.regression.unity.bots/Tests/Scripts/RegressionGames.Tests.asmdef.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 5ae8a2842b2149f383bdd81a625ce9c7
+timeCreated: 1726755434
